# api/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Observability & AIOps Platform
  version: 0.1.0
  description: Platform API. Telemetry stays in the backends. This API serves catalog, SLO results, incidents, and demo faults.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/demo/faults:
    get:
      summary: Active demo faults and the catalog
      responses:
        "200":
          description: Faults
        "404":
          description: Demo disabled
    post:
      summary: Enable a catalog fault
      responses:
        "200":
          description: Enabled
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []